Refactor item removal method to be more useful in determination.
Previously, it would return true if the sufficient amount of items was
removed, false otherwise. This is not comprehensive as there are 3 main
outcomes: if the full amount of items was removed, if an insufficient
amount of items was removed, or if the itemstack was invalid. The return
type is now an integer to appropriately accommodate this.