File logger will now append to existing file.
The file logger would, before, simply remove the old file and create a
new one with the same name every time the plugin was initialized or
reloaded. This is not useful as old logs will be lost, making the entire
feature useless, as users would have to search through the server logs
for information, which is what we attempted to avoid in the first place.

The file logger is now closed appropriately to avoid memory leaks.

Main class now documented much better